ready to submit

+++ src/get_list.c
@@ -6,7 +6,7 @@
 /*   By: brfialho <brfialho@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/13 19:03:16 by brfialho          #+#    #+#             */
-/*   Updated: 2025/10/13 20:42:52 by brfialho         ###   ########.fr       */
+/*   Updated: 2025/10/13 22:08:13 by brfialho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 t_list	**get_list(int argc, char *argv[])
 {
 	t_list	**head;
-	int		i;
+	long	i;
 
 	if (argc < 2)
 		exit(1);
